requirements.txt
numpy~=1.26.4
python-telegram-bot==21.1.1
